pyproject.toml
[project]
name = "dro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bitsandbytes>=0.45.5",
    "hydra-core>=1.3.2",
    "peft>=0.15.2",
    "transformers>=4.51.3",
    "trl>=0.16.1",
    "wandb>=0.19.9",
]
